backend/requirements.txt
Django==6.0.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-dotenv==1.0.0
dj-database-url==2.1.0
requests==2.31.0
cloudinary==1.36.0
djangorestframework-simplejwt==5.5.1
